Setting up the building and sim scripts for post-synthesis
Previously, we planned to move the PS output file to the rtl directory and
simulate it instead of the RTL design file. However, this was not a good
idea since we had to write new scripts to include ST, LU libraries and
SDF file back-annotation. So, moved the PS file back into the synthesis folder
and made the PS building and simulation scripts point to the the correct
file.

Makefiles has premade directive for PS simulation so utilized them in order
to have a clean vsim flow. vcompile_ps compiles the netlist file, ST/LU libs,
and the testbench only, avoiding compilation of IPs and all other components.

vsim_ps uses the PS libraries and SDF file